Search Component Implemented
